pyproject.toml
[project]
name = "sonar"
version = "0.1.0"
description = "Generate structured learning reports from articles with staged LLM orchestration."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.11.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.22",
    "trafilatura>=2.0.0",
    "pymupdf>=1.25.0",
]

[project.optional-dependencies]
local-classifier = [
    "transformers>=5.3.0",
    "torch>=2.10.0",
    "huggingface-hub>=1.6.0",
]

[project.scripts]
sonar = "main:main"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "ruff>=0.12.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["."]
